pyproject.toml
[tool.ruff]
target-version = "py311"
fix = true
line-length = 79
select = ["E", "W", "F", "N", "B", "UP", "I"]
ignore = ["E501"]

[tool.setuptools]
package-dir = {"" = "src"}
py-modules = ["main"]

[project]
name = "contact-assistant-bot"
description = "A contact management assistant bot"
authors = [
    { name = "Vladyslav" },
    { name = "Oleksandr" },
    { name = "Pavlo" },
    { name = "Olga" }
]
version = "0.1.0"
dependencies = [
    "colorama==0.4.6",
    "python-dotenv==1.1.1",
    "RapidFuzz==3.13.0",
    "tabulate==0.9.0",
    "ruff==0.12.5"
]

[project.scripts]
contact-bot = "main:main"